pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0"]

[project]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "pydantic>=2.9.0",
  "aiosqlite>=0.20.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "python-multipart>=0.0.20",
  "loguru>=0.7.0",
  "tqdm>=4.66.0",
  "pytest>=8.4.2",
  "pytest-asyncio>=0.24.0",
  "jinja2>=3.1.6",
  "pyyaml>=6.0.3",
  "litellm>=1.78.5",
  "rouge-score>=0.1.2",
  "scikit-learn>=1.3.0",
  "llama-index-core>=0.14.7",
  "anthropic>=0.73.0",
  "google-generativeai>=0.8.5",
  "ragas>=0.4.0",
  "pytest-timeout>=2.4.0",
  "langfuse==2.59.7",
  "instructor",
  "watchdog>=6.0.0",
  "typer>=0.9.0",
  "rich>=13.0.0",
]
description = "Q&A dataset generation and validation tool"
name = "datagenflow"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1.0"

[project.scripts]
dgf = "lib.cli.main:app"

[tool.setuptools.packages.find]
exclude = ["ui*", "data*"]
include = ["lib*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
ignore = []
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E501"]

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["scripts/"]

[[tool.mypy.overrides]]
disable_error_code = ["no-untyped-def", "no-untyped-call", "var-annotated", "override", "union-attr", "arg-type", "index", "type-arg", "unused-ignore", "import-not-found", "no-redef"]
module = "tests.*"

[[tool.mypy.overrides]]
disable_error_code = ["attr-defined", "no-untyped-def", "union-attr", "arg-type", "import-untyped"]
module = "debug_pipeline"

[[tool.mypy.overrides]]
disable_error_code = ["attr-defined", "no-untyped-def"]
module = "app"

[[tool.mypy.overrides]]
disable_error_code = ["no-any-return"]
module = "lib.storage"

[[tool.mypy.overrides]]
disable_error_code = ["import-untyped"]
module = "lib.blocks.builtin.langfuse"

[[tool.mypy.overrides]]
disable_error_code = ["import-untyped"]
module = "lib.blocks.builtin.ragas_metrics"

[tool.pytest.ini_options]
addopts = "-m 'not integration' --tb=short --ignore=scripts/ --ignore=tests/e2e/"
asyncio_mode = "strict"
markers = [
  "integration: integration tests requiring external services (ollama, etc) - only run when explicitly called",
]

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "playwright>=1.57.0",
]
